Complete Question 2
Implemented the typeclass Sized t that has one function, size, that is
type t -> Int, which returns the size of an object. Implemented Sized
for [a] where the size of the list is its length. Implemented Sized for
TwoFourTree a, where the size of the 2-4 tree is the number of data
elements it contains.