Complete Question 3
Implemented the abstract base class for TwoFourTree that is generic over
its element type and has an abstract virtual method called size, which takes
no arguments, returns an int, and is const in its implicit this parameter.
Implemented subclasses for each of the node types, TwoNode, ThreeNode and
FourNode. Each class has private fields of type T for the data elements,
along with private fields for TwoFourTrees. Each class has two constructors,
one for creating Nil nodes for leaves, another is for creating TwoFourTree
nodes for leaves from trees that already exist. Each class implements the
superclass size method, which returns the number of data elements in the the
tree rooted at that node.